Removes QSocketNotifier for the write end of the ZMQ Socket
It makes no sense to have a QSocketNotifier for the write end of the ZMQ
socket. Most sockets are available for writing most of the time. With no
way to disable this notifier from the public interface for the socket,
the socket notifier constantly emits the signal that its available for
writing resulting in stealing of 100% of CPU cycles.